autoLancer
I wrote this to reuse on freelancing sites that don't let you bulk
upload your portfolio details. https://codeshare.io/5cI9s

This is legacy js to be run in a tampermonkey script or a browser
console to hook into angularjs portfolio form on upwork.com

I used recursion instead of a for loop to delay the scripts execution
speed.. The plan is to port this script across many freelancing sites.